Fix invalid file paths in export_reports
Forward slashes ("/") are now replaced with underscores ("_") to prevent invalid file paths in Windows. The macro now checks if reports are saved and reports the accurate number of saved reports in the "macro complete" message at the end. Additionally, a small error of trying to trim the extension from the piece name (which has no extension) has been fixed.